+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Replaced use of Pillow image with numpy arrays, to match changes in wsidicom.
+
 ## [0.28.1] - 2026-07-10
 
 ### Fixed